+++ Editor/Settings/UniqueGameObjectsSettings.cs
@@ -1,0 +1,10 @@
+using GameCreator.Runtime.Common;
+
+namespace GameCreator.Editor.UniqueGameObject
+{
+    public class UniqueGameObjectsSettings : AssetRepository<UniqueGameObjectsRepository>
+    {
+        public override IIcon Icon => new IconID(ColorTheme.Type.TextLight);
+        public override string Name => "Unique Game Objects";
+    }
+}